Use `\@date` in header
Due to incompetence in my programming, `\mladate` had been
outputting the job date and not the document date.
Instead of fiddling with the code, I went with changing the
header to use the contents of `\@date` directly.  The only caveat
to this is that the date must be formatted correctly in the first
place by the user.

This fixes bug #2.